requirements.txt
Django==4.0
flake8>=3.9
black>=21.9b0
pytest-django
